add .half() to reduce memory usage

+++ finetune.py
@@ -90,7 +90,7 @@ def main(num_epochs = 10, lr=0.00002):
     device = torch.device("cuda")
 
     tokenizer = AutoTokenizer.from_pretrained("THUDM/chatglm2-6b", trust_remote_code=True)
-    model = AutoModel.from_pretrained("THUDM/chatglm2-6b", trust_remote_code=True).cuda() #do .half() for inference
+    model = AutoModel.from_pretrained("THUDM/chatglm2-6b", trust_remote_code=True).half().cuda() #do .half() for inference
     thawed_params = freezer(model, 3)
     model.train()
